refactor(analyzer): extract cyclomatic-complexity estimator into call-graph-complexity.ts (slice 3c)
Next safe slice of modularize-call-graph-builder. Moves the CYCLOMATIC COMPLEXITY
section (computeCyclomaticComplexity + the CC_PATTERN_* regex tables) out of
call-graph.ts into call-graph-complexity.ts — a pure, dependency-free McCabe estimator.

computeCyclomaticComplexity was exported from call-graph.ts (though no external
importer today), so it is imported back AND re-exported on the barrel to preserve
the public surface exactly; the CC_PATTERN_* tables stay private to the new module.
call-graph.ts: 4,887 → 4,879 lines.

Behavior-preserving, verified four ways:
- Snapshot oracle first extended to exercise cyclomaticComplexity > 1 (branchy TS +
  Python fixtures: risky cc=5, handler cc=3), then captured before/after: IDENTICAL —
  SHA-256 7b765f31…
- Public export surface byte-identical (0 added, 0 removed; computeCyclomaticComplexity
  still exported).
- tsc compiles all importers clean; build/lint/typecheck green.
- Full suite green: 279 files / 5534 tests.

No feature, no dependency, no LLM, no persisted artifact.
